Many changes to the way Inputs and Outputs are handled
This change removes Force Feedback from the Fingers and into its own
output. This change also adds Clamping FFB using a digital pin and it
updates the Output interface to allow outputs to do work even if they
haven't recieved an update from the driver this iteration of the loop.

Finally, we also move the hardware config to its own file to clean up
the main ino file.